Renamed id_t to lwk_id_t and added user hash functions. - Global search and replace on id_t to lwk_id_t (avoid clashes in <sys/types.h>) - Extended id_t to 64-bits - Added hash function argument to htable_create to allow non-integer hash tables

+++ include/lwk/hash.h
@@ -65,6 +65,6 @@ static inline u32 hash_32(u32 val, unsigned int bits)
 
 static inline unsigned long hash_ptr(void *ptr, unsigned int bits)
 {
-	return hash_long((unsigned long)ptr, bits);
+	return hash_long((uintptr_t)ptr, bits);
 }
 #endif /* _LINUX_HASH_H */
